move itemAdditionalTimesInfo & getParticipantTypeForItemWithDurationManagedByUser() into api/contests/get_members_additional_times.go
