+++ app/api/contests/get_members_additional_times.go
@@ -6,8 +6,23 @@ import (
 	"github.com/go-chi/render"
 	"github.com/jinzhu/gorm"
 
+	"github.com/France-ioi/AlgoreaBackend/v2/app/database"
 	"github.com/France-ioi/AlgoreaBackend/v2/app/service"
 )
+
+// swagger:model itemAdditionalTimesInfo
+type itemAdditionalTimesInfo struct {
+	// required: true
+	GroupID int64 `json:"group_id,string"`
+	// required: true
+	Name string `json:"name"`
+	// required: true
+	Type string `json:"type"`
+	// required: true
+	AdditionalTime int32 `json:"additional_time"`
+	// required: true
+	TotalAdditionalTime int32 `json:"total_additional_time"`
+}
 
 // swagger:operation GET /items/{item_id}/groups/{group_id}/members/additional-times items itemListMembersAdditionalTime
 //
@@ -161,4 +176,13 @@ func (srv *Service) getMembersAdditionalTimes(w http.ResponseWriter, r *http.Req
 
 	render.Respond(w, r, result)
 	return service.NoError
+}
+
+func getParticipantTypeForItemWithDurationManagedByUser(
+	store *database.DataStore, itemID int64, user *database.User,
+) (string, error) {
+	var participantType string
+	err := store.Items().WithDurationByIDAndManagedByUser(itemID, user).
+		PluckFirst("items.entry_participant_type", &participantType).Error()
+	return participantType, err
 }
